Explicitly inject widget script into document
